+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2021-12-24
 ### Added
 - add `https://www.youtube.com/`
+- add items related to LibreOffice to the launcher
 
 ### Fixed
 - fix `rm: missing operand`